refactor : database 설정 관련 클래스 database.config 로 패키지 이동
